renamed main function file
